# specs/design/api-contracts.schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OnboardX API",
    "version": "1.0.0",
    "description": "Customer onboarding and KYC platform (synthetic data). Integer-only numerics. See specs/design/api-contracts.md."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "local verification mode"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/reports/tat": {
      "get": {
        "operationId": "report_tat",
        "summary": "Report tat",
        "tags": [
          "admin"
        ],
        "x-story": "E5-S3 (E5-S1)",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TatReport"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "product",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Product"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ]
      }
    },
    "/api/v1/admin/reports/funnel": {
      "get": {
        "operationId": "report_funnel",
        "summary": "Report funnel",
        "tags": [
          "admin"
        ],
        "x-story": "E5-S3 (E5-S2)",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnelReport"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "product",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Product"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ]
      }
    },
    "/api/v1/admin/reports/backlog": {
      "get": {
        "operationId": "report_backlog",
        "summary": "Report backlog",
        "tags": [
          "admin"
        ],
        "x-story": "E5-S3 (E5-S2)",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacklogReport"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "product",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Product"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ]
      }
    },
    "/api/v1/admin/reports/time-per-stage": {
      "get": {
        "operationId": "report_time_per_stage",
        "summary": "Report time-per-stage",
        "tags": [
          "admin"
        ],
        "x-story": "E5-S3 (E5-S1)",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimePerStageReport"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "product",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Product"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ]
      }
    },
    "/api/v1/admin/reports/rejection-reasons": {
      "get": {
        "operationId": "report_rejection_reasons",
        "summary": "Report rejection-reasons",
        "tags": [
          "admin"
        ],
        "x-story": "E5-S3 (E5-S2)",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectionReasonsReport"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "product",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Product"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ]
      }
    },
    "/api/v1/admin/reports/auto-approval": {
      "get": {
        "operationId": "report_auto_approval",
        "summary": "Report auto-approval",
        "tags": [
          "admin"
        ],
        "x-story": "E5-S3 (E5-S2)",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoApprovalReport"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "product",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Product"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Product": {
        "type": "string",
        "enum": [
          "Savings",
          "Current",
          "NRE"
        ]
      },
      "CaseState": {
        "type": "string",
        "enum": [
          "INITIATED",
          "DOCS_SUBMITTED",
          "SCREENED",
          "CLASSIFIED",
          "APPROVED",
          "REJECTED",
          "MANUAL_REVIEW"
        ]
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNAUTHENTICATED",
          "FORBIDDEN",
          "NOT_FOUND",
          "INVALID_STATE",
          "CASE_LOCKED",
          "PROFILE_LOCKED",
          "RULESET_IMMUTABLE",
          "FILE_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "VALIDATION_ERROR",
          "UNKNOWN_CHECKLIST_ITEM",
          "MISSING_DOCUMENTS",
          "MISSING_PROFILE",
          "MISSING_PROFILE_FIELD",
          "RULESET_INVALID",
          "UNKNOWN_REASON_CODE",
          "DRAFT_EXISTS",
          "ALREADY_DEACTIVATED"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "error"
        ]
      },
...
      "ReportFilters": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product",
            "nullable": true
          },
          "from": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "to": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "required": [
          "product",
          "from",
          "to"
        ]
      },
      "TatReport": {
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ReportFilters"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "product": {
                  "$ref": "#/components/schemas/Product"
                },
                "count": {
                  "type": "integer"
                },
                "avg_seconds": {
                  "type": "integer"
                },
                "min_seconds": {
                  "type": "integer"
                },
                "max_seconds": {
                  "type": "integer"
                }
              },
              "additionalProperties": false,
              "required": [
                "product",
                "count",
                "avg_seconds",
                "min_seconds",
                "max_seconds"
              ]
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "filters",
          "items"
        ]
      },
      "FunnelReport": {
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ReportFilters"
          },
          "stages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "stage": {
                  "type": "string",
                  "enum": [
                    "INITIATED",
                    "DOCS_SUBMITTED",
                    "SCREENED",
                    "CLASSIFIED",
                    "MANUAL_REVIEW",
                    "APPROVED",
                    "REJECTED"
                  ]
                },
                "count": {
                  "type": "integer"
                },
                "conversion_bp": {
                  "type": "integer"
                }
              },
              "additionalProperties": false,
              "required": [
                "stage",
                "count",
                "conversion_bp"
              ]
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "filters",
          "stages"
        ]
      },
      "BacklogReport": {
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ReportFilters"
          },
          "count": {
            "type": "integer"
          },
          "oldest_age_minutes": {
            "type": "integer"
          },
          "buckets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string",
                  "enum": [
                    "under_60",
                    "60_to_1440",
                    "over_1440"
                  ]
                },
                "min_minutes": {
                  "type": "integer"
                },
                "max_minutes": {
                  "type": "integer",
                  "nullable": true
                },
                "count": {
                  "type": "integer"
                }
              },
              "additionalProperties": false,
              "required": [
                "label",
                "min_minutes",
                "max_minutes",
                "count"
              ]
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "filters",
          "count",
          "oldest_age_minutes",
          "buckets"
        ]
      },
      "TimePerStageReport": {
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ReportFilters"
          },
          "stages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "from_state": {
                  "$ref": "#/components/schemas/CaseState"
                },
                "to_state": {
                  "$ref": "#/components/schemas/CaseState"
                },
                "count": {
                  "type": "integer"
                },
                "avg_seconds": {
                  "type": "integer"
                }
              },
              "additionalProperties": false,
              "required": [
                "from_state",
                "to_state",
                "count",
                "avg_seconds"
              ]
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "filters",
          "stages"
        ]
      },
      "RejectionReasonsReport": {
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ReportFilters"
          },
          "items": {
            "type": "array",
            "maxItems": 5,
            "items": {
              "type": "object",
              "properties": {
                "reason_code": {
                  "type": "string"
                },
                "count": {
                  "type": "integer"
                }
              },
              "additionalProperties": false,
              "required": [
                "reason_code",
                "count"
              ]
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "filters",
          "items"
        ]
      },
      "AutoApprovalReport": {
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ReportFilters"
          },
          "auto_approved": {
            "type": "integer"
          },
          "decided": {
            "type": "integer"
          },
          "rate_bp": {
            "type": "integer"
          },
          "target_bp": {
            "type": "integer",
            "enum": [
              6000
            ]
          },
          "met": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "required": [
          "filters",
          "auto_approved",
          "decided",
          "rate_bp",
          "target_bp",
          "met"
        ]
      }
    }
  }
}